refactor: #220 - Header 컴포넌트 useHeader 훅 분리
