+++ src/components/common/Header/Header.tsx
@@ -1,53 +1,24 @@
 'use client'
 
-import { useCallback, useEffect, useState } from 'react'
 import { LinkIcon } from '@heroicons/react/20/solid'
 import { BellIcon } from '@heroicons/react/24/outline'
 import { MagnifyingGlassCircleIcon } from '@heroicons/react/24/outline'
 import { Bars3Icon } from '@heroicons/react/24/solid'
 import Link from 'next/link'
-import { usePathname, useRouter, useSearchParams } from 'next/navigation'
 import Button from '../Button/Button'
 import SearchModal from '../SearchModal/SearchModal'
 import Sidebar from '../Sidebar/Sidebar'
+import useHeader from './hooks/useHeader'
 
 const Header = () => {
-  const pathname = usePathname()
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const [isSidebarOpen, setIsSidebarOpen] = useState(false)
-  const isSearchModalOpen = searchParams.get('search')
-  const currentPage = pathname
-    .split(/[^a-zA-Z]/)[1] // 라우터명
-    .replace(/^[a-z]/, (char) => char.toUpperCase()) // 첫글자 대문자 치환
-
-  const createQueryString = useCallback(
-    (name: string, value: string) => {
-      const params = new URLSearchParams(searchParams)
-      params.set(name, value)
-
-      return params.toString()
-    },
-    [searchParams],
-  )
-
-  const deleteQueryString = useCallback(
-    (name: string) => {
-      const params = new URLSearchParams(searchParams)
-      params.delete(name)
-
-      return params.toString()
-    },
-    [searchParams],
-  )
-
-  useEffect(() => {
-    if (isSidebarOpen || isSearchModalOpen) {
-      document.body.style.overflow = 'hidden'
-    } else {
-      document.body.style.overflow = 'auto'
-    }
-  }, [isSidebarOpen, isSearchModalOpen])
+  const {
+    currentPage,
+    isSidebarOpen,
+    isSearchModalOpen,
+    openSearchModal,
+    closeSearchModal,
+    setIsSidebarOpen,
+  } = useHeader()
 
   return (
     <>
@@ -70,12 +41,7 @@ const Header = () => {
           </Button>
           <Button
             className="flex h-8 w-8 items-center justify-center"
-            onClick={() =>
-              router.replace(
-                pathname + '?' + createQueryString('search', 'true'),
-                { scroll: false },
-              )
-            }>
+            onClick={openSearchModal}>
             <MagnifyingGlassCircleIcon className="h-6 w-6 text-slate9" />
           </Button>
           <Button
@@ -91,15 +57,7 @@ const Header = () => {
           onClose={() => setIsSidebarOpen(false)}
         />
       )}
-      {isSearchModalOpen && (
-        <SearchModal
-          onClose={() =>
-            router.replace(pathname + '?' + deleteQueryString('search'), {
-              scroll: false,
-            })
-          }
-        />
-      )}
+      {isSearchModalOpen && <SearchModal onClose={closeSearchModal} />}
     </>
   )
 }
